requirements.txt
Flask>=3.0.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
stripe>=11.0.0,<12.0.0
gunicorn>=22.0.0,<23.0.0
psycopg[binary]>=3.2.0,<4.0.0
